refactor(org-events): replace attend endpoint with view-as staff rosters
Rename the admin-only attend endpoint to view-as. It now upserts a hidden
is_staff roster keyed by (event, user, type) instead of a participant row, so
admins can use coach/dancer features (which need a real roster id) without
polluting participant data. No participant flow used attend — it was always
admin-gated.

OrgEventMiddleware resolves the correct sandbox via the x-act-as-type request
header when an admin holds both a coach and a dancer staff roster; real users
have a single roster, so behaviour is unchanged for them.
