+++ apps/backend/app/modules/orgs/events/routes.ts
@@ -17,8 +17,7 @@ const FiltersRosterController = () => import("./rosters/filters/controller.ts");
 const StatsRosterController = () => import("./rosters/stats/controller.ts");
 const ResendInvitesController = () =>
   import("./rosters/resend-invites/controller.ts");
-const AttachAccountController = () =>
-  import("./rosters/attach/controller.ts");
+const AttachAccountController = () => import("./rosters/attach/controller.ts");
 const SearchDancerUsersController = () =>
   import("./rosters/attach/search-controller.ts");
 const ListChecklistController = () => import("./checklist/list/controller.ts");
@@ -54,12 +53,10 @@ const UpdateVideoController = () => import("./videos/update/controller.ts");
 const DeleteVideoController = () => import("./videos/delete/controller.ts");
 const AudioUploadUrlController = () =>
   import("./videos/audio-upload-url/controller.ts");
-const AttendEventController = () => import("./attend/controller.ts");
+const ViewAsController = () => import("./view-as/controller.ts");
 const CheckInController = () => import("./check-in/controller.ts");
-const AdminCheckInController = () =>
-  import("./rosters/check-in/controller.ts");
-const CheckInStatusController = () =>
-  import("./check-in/status-controller.ts");
+const AdminCheckInController = () => import("./rosters/check-in/controller.ts");
+const CheckInStatusController = () => import("./check-in/status-controller.ts");
 const ScheduleController = () => import("./schedule/controller.ts");
 
 router
@@ -308,7 +305,7 @@ router
         middleware.orgAdmin(),
       ]);
     router
-      .post(":slug/events/attend", [AttendEventController])
+      .post(":slug/events/view-as", [ViewAsController])
       .use([
         middleware.auth(),
         middleware.org(),
@@ -318,11 +315,7 @@ router
       ]);
     router
       .get(":slug/events/:id/check-in/status", [CheckInStatusController])
-      .use([
-        middleware.auth(),
-        middleware.org(),
-        middleware.orgMember(),
-      ]);
+      .use([middleware.auth(), middleware.org(), middleware.orgMember()]);
     router
       .post(":slug/events/:id/check-in", [CheckInController])
       .use([
@@ -343,22 +336,12 @@ router
       ]);
     router
       .get(":slug/events/:id/schedule", [ScheduleController])
-      .use([
-        middleware.auth(),
-        middleware.org(),
-        middleware.orgMember(),
-      ]);
+      .use([middleware.auth(), middleware.org(), middleware.orgMember()]);
 
     // Video categories
     router
-      .get(":slug/events/:id/video-categories", [
-        ListVideoCategoriesController,
-      ])
-      .use([
-        middleware.auth(),
-        middleware.org(),
-        middleware.orgMember(),
-      ]);
+      .get(":slug/events/:id/video-categories", [ListVideoCategoriesController])
+      .use([middleware.auth(), middleware.org(), middleware.orgMember()]);
     router
       .post(":slug/events/:id/video-categories", [
         CreateVideoCategoryController,
@@ -383,11 +366,7 @@ router
     // Videos
     router
       .get(":slug/events/:id/videos", [ListVideosController])
-      .use([
-        middleware.auth(),
-        middleware.org(),
-        middleware.orgMember(),
-      ]);
+      .use([middleware.auth(), middleware.org(), middleware.orgMember()]);
     router
       .post(":slug/events/:id/videos", [CreateVideoController])
       .use([
